move firebase.js to firebase directory & update signInWithEmailAndPassword
